Add files via upload
The project involves a website known as wykop, a polish page which brings together various news from the internet. It was created for one of my classes with a colleague.
It's aim was to present specific information found, such as trendings and hot topics being disscused, containing theirs statistics, such as upvote scores.